tutorials/08_discord/requirements.txt
python-dotenv==0.13.0
discord==1.0.1
discord.py[voice]==1.3.3
